+++ src/cache_entry.h
@@ -92,9 +92,9 @@ public:
     size_t                  size() const;
     /**
      * @brief Get the video frame count.
-     * @return On success, returns the number of frames; on error, returns AV_NOPTS_VALUE (calculation failed or no video source file).
+     * @return On success, returns the number of frames; on error, returns 0 (calculation failed or no video source file).
      */
-    int64_t                 video_frame_count() const;
+    uint32_t                video_frame_count() const;
     /**
      * @brief Get the age of the cache entry.
      * @return Returns the age of the cache entry in seconds since epoch.
